Change Rank number 0 = root 255 = no rank TODO: add verification to not set rank to 255 with increment

+++ src/Common/PacketType.c
@@ -17,7 +17,7 @@ struct abstract_packet
 struct general_packet 
 {
     uint8_t type;
-    uint16_t rank;
+    uint8_t rank;
 };
 
 // TODO enhance
